feat: add support to return only one result per matching line
The --vimgrep option normally used to get results in a format compatible
with the quickfix list returns one result for each match on each
line. This means that the same line may appear many times in the
results.

Sometimes this is not desirable, so support is added to use the
`--line-number` option instead, in the underlying command. This can be
set globally, or on an individual search by simply using the native
ripgrep flags, `-n` or `--line-number`.